+++ CHANGELOG.md
@@ -1,6 +1,15 @@
+## 0.5.6 (December 18th, 2025)
+### IMPROVEMENTS:
+* Added the optional setting of 'log_level', which is used to control console messages from salt-call.
+
+## 0.5.5 (October 29th, 2025)
+### IMPROVEMENTS:
+* The default OS remains "linux", but the provisioner now detects "windows" automatically.
+* The target_os option is to be deprecated in a future release.
+* Go version updated to 1.25
+
 # 0.5.1 (October 1st, 2025)
 ### BUGFIXES:
-
 * Windows environment variables are formatted correctly
 * Hardcoded linux command removed from create directory function (naughty developer)
 * Hardcoded linux command removed from delete directory function (ooops)